+++ src/index.d.ts
@@ -140,7 +140,7 @@ export declare class Account {
   signIn(data: AuthenticateData): Promise<AxiosResponse<CreateAccount>>;
   signUp(data: SignUpData): Promise<AxiosResponse<CreateAccount>>;
   signOut(): Promise<AxiosResponse<undefined>>;
-  refreshToken(clientId: number | string): Promise<AxiosResponse<CreateAccount>>;
+  refreshToken(clientId: string): Promise<AxiosResponse<CreateAccount>>;
   reportSSOtoken(
     ssoDomain: string,
     token: string,